+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 - Drop support for PHP < 7.3 [#119](https://github.com/rlanvin/php-rrule/issues/119)
 - Add support for PHP 8.4
 
+### Fixed
+
+- Added a workaround for a Carbon 3 bug that makes occursAt fail in some cases [#164](https://github.com/rlanvin/php-rrule/issues/164)
+
 ## [2.6.0] - 2025-04-25
 
 ### Added